# openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "SplitWiseMoney API",
                 "description":  "SplitWiseMoney - Expense sharing web application API documentation",
                 "version":  "1.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
...
    "paths":  {
...
                  "/api/groups/{id}":  {
                                           "get":  {
                                                       "tags":  [
                                                                    "Groups"
                                                                ],
                                                       "summary":  "Get specific group details by ID",
                                                       "operationId":  "getGroupById",
                                                       "parameters":  [
                                                                          {
                                                                              "name":  "id",
                                                                              "in":  "path",
                                                                              "required":  true,
                                                                              "schema":  {
                                                                                             "type":  "integer",
                                                                                             "format":  "int64"
                                                                                         }
                                                                          }
                                                                      ],
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "OK",
                                                                                     "content":  {
                                                                                                     "*/*":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/GroupResponse"
                                                                                                                            }
                                                                                                             }
                                                                                                 }
                                                                                 }
                                                                     },
                                                       "security":  [
                                                                        {
                                                                            "bearerAuth":  [

                                                                                           ]
                                                                        }
                                                                    ]
                                                   },
                                           "put":  {
                                                       "tags":  [
                                                                    "Groups"
                                                                ],
                                                       "summary":  "Update group name",
                                                       "operationId":  "editGroup",
                                                       "parameters":  [
                                                                          {
                                                                              "name":  "id",
                                                                              "in":  "path",
                                                                              "required":  true,
                                                                              "schema":  {
                                                                                             "type":  "integer",
                                                                                             "format":  "int64"
                                                                                         }
                                                                          }
                                                                      ],
                                                       "requestBody":  {
                                                                           "content":  {
                                                                                           "application/json":  {
                                                                                                                    "schema":  {
                                                                                                                                   "$ref":  "#/components/schemas/GroupRequest"
                                                                                                                               }
                                                                                                                }
                                                                                       },
                                                                           "required":  true
                                                                       },
                                                       "responses":  {
                                                                         "200":  {
                                                                                     "description":  "OK",
                                                                                     "content":  {
                                                                                                     "*/*":  {
                                                                                                                 "schema":  {
                                                                                                                                "$ref":  "#/components/schemas/GroupResponse"
                                                                                                                            }
                                                                                                             }
                                                                                                 }
                                                                                 }
                                                                     },
                                                       "security":  [
                                                                        {
                                                                            "bearerAuth":  [

                                                                                           ]
                                                                        }
                                                                    ]
                                                   },
                                           "delete":  {
                                                          "tags":  [
                                                                       "Groups"
                                                                   ],
                                                          "summary":  "Delete group",
                                                          "operationId":  "deleteGroup",
                                                          "parameters":  [
                                                                             {
                                                                                 "name":  "id",
                                                                                 "in":  "path",
                                                                                 "required":  true,
                                                                                 "schema":  {
                                                                                                "type":  "integer",
                                                                                                "format":  "int64"
                                                                                            }
                                                                             }
                                                                         ],
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "OK",
                                                                                        "content":  {
                                                                                                        "*/*":  {
                                                                                                                    "schema":  {
                                                                                                                                   "type":  "string"
                                                                                                                               }
                                                                                                                }
                                                                                                    }
                                                                                    }
                                                                        },
                                                          "security":  [
                                                                           {
                                                                               "bearerAuth":  [

                                                                                              ]
                                                                           }
                                                                       ]
                                                      }
                                       },
...
                  "/api/settlements/group/{groupId}":  {
                                                           "get":  {
                                                                       "tags":  [
                                                                                    "Settlements"
                                                                                ],
                                                                       "summary":  "List all recorded settlement transactions in a group",
                                                                       "operationId":  "getGroupSettlements",
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "groupId",
                                                                                              "in":  "path",
                                                                                              "required":  true,
                                                                                              "schema":  {
                                                                                                             "type":  "integer",
                                                                                                             "format":  "int64"
                                                                                                         }
                                                                                          }
                                                                                      ],
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "OK",
                                                                                                     "content":  {
                                                                                                                     "*/*":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "type":  "array",
                                                                                                                                                "items":  {
                                                                                                                                                              "$ref":  "#/components/schemas/SettlementResponse"
                                                                                                                                                          }
                                                                                                                                            }
                                                                                                                             }
                                                                                                                 }
                                                                                                 }
                                                                                     },
                                                                       "security":  [
                                                                                        {
                                                                                            "bearerAuth":  [

                                                                                                           ]
                                                                                        }
                                                                                    ]
                                                                   },
                                                           "post":  {
                                                                        "tags":  [
                                                                                     "Settlements"
                                                                                 ],
                                                                        "summary":  "Record a settlement payment between group members",
                                                                        "operationId":  "createSettlement",
                                                                        "parameters":  [
                                                                                           {
                                                                                               "name":  "groupId",
                                                                                               "in":  "path",
                                                                                               "required":  true,
                                                                                               "schema":  {
                                                                                                              "type":  "integer",
                                                                                                              "format":  "int64"
                                                                                                          }
                                                                                           }
                                                                                       ],
                                                                        "requestBody":  {
                                                                                            "content":  {
                                                                                                            "application/json":  {
                                                                                                                                     "schema":  {
                                                                                                                                                    "$ref":  "#/components/schemas/SettlementRequest"
                                                                                                                                                }
                                                                                                                                 }
                                                                                                        },
                                                                                            "required":  true
                                                                                        },
                                                                        "responses":  {
                                                                                          "200":  {
                                                                                                      "description":  "OK",
                                                                                                      "content":  {
                                                                                                                      "*/*":  {
                                                                                                                                  "schema":  {
                                                                                                                                                 "$ref":  "#/components/schemas/SettlementResponse"
                                                                                                                                             }
                                                                                                                              }
                                                                                                                  }
                                                                                                  }
                                                                                      },
                                                                        "security":  [
                                                                                         {
                                                                                             "bearerAuth":  [

                                                                                                            ]
                                                                                         }
                                                                                     ]
                                                                    }
                                                       },
...
                  "/api/groups":  {
                                      "get":  {
                                                  "tags":  [
                                                               "Groups"
                                                           ],
                                                  "summary":  "List all groups associated with the current user",
                                                  "operationId":  "getUserGroups",
                                                  "responses":  {
                                                                    "200":  {
                                                                                "description":  "OK",
                                                                                "content":  {
                                                                                                "*/*":  {
                                                                                                            "schema":  {
                                                                                                                           "type":  "array",
                                                                                                                           "items":  {
                                                                                                                                         "$ref":  "#/components/schemas/GroupResponse"
                                                                                                                                     }
                                                                                                                       }
                                                                                                        }
                                                                                            }
                                                                            }
                                                                },
                                                  "security":  [
                                                                   {
                                                                       "bearerAuth":  [

                                                                                      ]
                                                                   }
                                                               ]
                                              },
                                      "post":  {
                                                   "tags":  [
                                                                "Groups"
                                                            ],
                                                   "summary":  "Create a new expense group",
                                                   "operationId":  "createGroup",
                                                   "requestBody":  {
                                                                       "content":  {
                                                                                       "application/json":  {
                                                                                                                "schema":  {
                                                                                                                               "$ref":  "#/components/schemas/GroupRequest"
                                                                                                                           }
                                                                                                            }
                                                                                   },
                                                                       "required":  true
                                                                   },
                                                   "responses":  {
                                                                     "200":  {
                                                                                 "description":  "OK",
                                                                                 "content":  {
                                                                                                 "*/*":  {
                                                                                                             "schema":  {
                                                                                                                            "$ref":  "#/components/schemas/GroupResponse"
                                                                                                                        }
                                                                                                         }
                                                                                             }
                                                                             }
                                                                 },
                                                   "security":  [
                                                                    {
                                                                        "bearerAuth":  [

                                                                                       ]
                                                                    }
                                                                ]
                                               }
                                  },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "GroupRequest":  {
                                                            "type":  "object",
                                                            "properties":  {
                                                                               "groupName":  {
                                                                                                 "type":  "string",
                                                                                                 "maxLength":  100,
                                                                                                 "minLength":  0
                                                                                             }
                                                                           },
                                                            "required":  [
                                                                             "groupName"
                                                                         ]
                                                        },
                                       "GroupResponse":  {
                                                             "type":  "object",
                                                             "properties":  {
                                                                                "id":  {
                                                                                           "type":  "integer",
                                                                                           "format":  "int64"
                                                                                       },
                                                                                "groupName":  {
                                                                                                  "type":  "string"
                                                                                              },
                                                                                "createdById":  {
                                                                                                    "type":  "integer",
                                                                                                    "format":  "int64"
                                                                                                },
                                                                                "createdByName":  {
                                                                                                      "type":  "string"
                                                                                                  },
                                                                                "createdAt":  {
                                                                                                  "type":  "string",
                                                                                                  "format":  "date-time"
                                                                                              }
                                                                            }
                                                         },
...
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "name":  "bearerAuth",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}